dist-server/requirements.txt
fastapi>=0.123.5
grpcio>=1.76.0
grpcio-tools>=1.76.0
requests>=2.32.5
uvicorn>=0.38.0
